Refactoring the Attribute initialization
Implemented the ApplyEffectToSelf function that does the same as InitializePrimaryAttributes but with an Input for UGameplayEffect and an float for the Level.
Adde InitializeDefaultAttributes that calls ApplyEffectToSelf with each the inputs for DefaultPrimaryAttributes and DefaultSecondaryAttributes.
DefaultSecondaryAttributes variable also added before in the Header file.
Than InitializeDefaultAttributes gets called in AuraCharacter::InitAbilityActorInfo()