Added GameplayAbilities
Looping through an TArray of UGameplayAbility in AuraAbilitySystemComponent in the AddCharacterAbilites function. Call the function in AuraCharacterBase in the AddCharacterAbilities function and checking if it HasAuthority. Finaly call the inherited func in the PossesedBy function in AuraCharacter. In Editor made a BP based on AuraGameplayAbility and added some debug strings for activate and ended. Added the BP to BP_AuraCharacter.