Setting Up Auto Running
With help of UNavigationSystemV1::FindPathToLocationSyncronously I can loop through all points inside an NavmeshBoundVolume from the ActorLocation to the CaachedDestination and add it to a Spline to get a smoother Path to counter snapping. Drawing debug sphere to check if it is working even with a Mesh in the path.