Implementing Auto Running
Added AutoRun function. With some Spline specific function I can get FindDirectionClosestToWorldLocation FVector. Set it into the AddMovementInput function. Also checking if the DistanceToDestination <= AutoRunAsseptanceRadius. If it is then no AutoRunning anymore. Also added for the ChachedDestination the LastIndex in the array so I get the last point that got generated. Helps if there is now last point ouside the navmeshvolume or in between assets. Will auto run forever otherwise. For the assets -> ingoring the visibility channel to cursor trace behind it.