Spawning Projectile
In the CombatInterface I created an function GetCombatSocketLocation to return an FVector.
In AuraCharacterBase I override it and also created an FName variable called WeoponTipSocketName as EditAnywhere.
In the function I call from the Weapon USkeletalMeshComp the SocketLocation that is Stored in the WeaponTipSocketName variable. In the BP_AuraCharacter set the correct name of the socket of the weapon SKM.
In AuraProjectile is set bReplicates = true;
In AuraProjectileSpell I get AAuraProjectile. In the AcitvateAbility I use the SpawnActorDeferred to spawn the projectile. And FinishSpawning to spawn the actor finally after setting the parameters for it.