Sending Gameplay Events
Added AN_MontageEvent BP based on AnimNotify and in the Project Settings added a new Gameplay Tag-> Event.Montage.FireBolt. Function call-> SendGameplayEventtoActor with an EventTag variable with the datatype GameplayTag. Calling this Notify in the AnimMontage for the FireBolt and in GA_FireBolt with the WaitGameplayEvent Checking for the Event.Montage.FireBolt event tag.