Receiving Target Data
Now the server is getting the Data. So if we are not locally controlled -> Server we are calling the AbilityTargetDataSetDelegate and with the OnTargetDataReplicatedCallback func we can broadcasting the DataHandle. In AuraAssetManager wie call InitGlobalData with UAbilitySystemGlobals. After that we can call in GA_FireBolt from the DataHanlde output GetHitResult from TargetData and break its containing data and get its location. With DrawDebugSphere we can check if it is working.