Added a AnimNotify BP
Added a AnimNotify BP based on AnimNotify class. In it I override the received notify function and call the func SendGameplayEventToActor. Getting the owner as the Actor and for the EventTag input a EventTag variable based on GameplayTag.
In the AnimMontage added the AN_MontageEvent in the right spot of the Animation.