Created TargetDataUnderMouse class
Created TargetDataUnderMouse class based on UAbilityTask class. Added a static function which is BlueprinCallable. Added also a Delegatte to have an ExecPin ValidData and Data that is holding the data what is put into it. Also added the Activate func that gets the CursorHit from the PlayerController and than broadcasting the Impact Point with the delegate.
In GA_FireBolt in the Output pin for ValidData I call the drawdebugsphere and for the location I connect the data pin to it.