Send Mouse Cursor Data
SendMouseCursorData() function making sure everything is predicted for the server.
The Delegate datatype is now FGameplayAbilityTargetDataHandle instead of FVector. We getting the Location now from the DataHandle because the it is getting the FHitResult Data. After that we call ServerSetReplicatedTargetData where also the DataHandle with now the FHitResult Data gets replicated and other stuff.