Orienting the Projectile
Added in the SpawnProjectile() function of AuraProjectileSpell an FRotator to calculate the rotation. Works as intended. Also as an input an FVector to plug the Location data from the TargetDataUnderMouse ->HitResult. !!EndAbility is not replicated-> On Client side the projectile will not show somehow if the EndAbility at the end is connected.
Added a new InputAction IA_SHIFT-> Shift-Key.
Added in AuraPlayerController::SetupInputComponent the BindAction function for  pressing and releasing the button with functions that return a bool.