MotionWarping
Plugin->MotionWarping. In BP_AuraCharacter added the Component. In the AM_Cast_FireBolt added the MotionWarping notify.
In CombatInterface.h added the func UpdateFacingTarget with an FVector input and added the UFUNCTION to be implementable as an Event and Callable as an Function.
In BP_AuraCharacter added the event and hooked the function AddOrUpdateWarpTargetFromLocation with the MotionWarpinComp as Target the the FVector as TargetLocation. In GA_FireBolt GetAvatarActorFromActorInfo->CastToCombatInterface and calling the UpdateFacingTarget function with the ProjectileTargetLocation as Target.