Projectile Impact
In AuraProjectile added some variables for the SoundBase and NiagaraSystem and an override for Destroyed(). OnOverlap an Explosion effect gets played along with a SoundEffect. Of if it gets destroyed. Whatever happens first. For an Sound that play as long the projectile is flying I added another variable ->USoundBase and an Component UAudioComponent. In BeginPlay Instead of PlaySoundAtLocation (the other sounds) I call SpawnSoundAttached and save it in the Component and just call the Stop() func in Overlap and Destroyed. In BP_FireBold I added the sfx Sounds and NiagaraEffect into the variables.