CharacterClassInfo
Added UCharacterClassInfo based on UDataAsset. It is holding an Enum for each CharacterClass. Also a struct with UGameplayEffect for the PrimarayAttributes. In the class UGameplayEffects for the secondaryAttributes and VitalAttributes. Also a TMap that for each enum CharacterClass I can choose an PrimaryEffect of the struct. So each class can have different primaryEffects.  In Editor created a DataAsset based on the UCharacterClassInfo class.