SetByCallerMagnitude
Added in AuraGameplayTags Damage as a new GameplayTag.
In AuraProjectileSpell::SpawnProjectile right before FinishSpawning we get the GameplayTags an call the UAbilitySystemBlueprintLibrary::AssingTagSetBayCallerMagnitude. So now we have a GameplayTag and an value for the damage. So now we set the damage amount in the GameplayAbility and GE_Damage will get the value that have the correct tag.