EnemyDeath
Added in CombatInterface the Die function to call it in AuraAttributeSet if the bFatal bool is true otherwise just HitReact. In AuraCharacterBase the Die func get overridden and also added MulitcastHandleDeath func and added NetMulticast Reliable so the behavior is replicated. In the Implementation setting the behavior for the weapon, the mesh and the capsulecomp. The weapon detachment is in Die func and the implementation get called. In AuraEnemy getting the Die func and calling Super and setting also the lifespan with an variable LifeSpan.