DissolveEffect
Added a DissolveEffect when an enemy dies. In the Dissolve func check if we set an MaterialInstance in the Editor. If so we create an DymaicMaterial so we can change it in an Timline. The StartDissolveTimeline is an BlueprintImplementableEvent. So we call it in BP_EnemyBase from there we call a Timline func and set a Track and call SetScalarParamterValue. So the Paramater in the DissolveMaterial Dissolve Value will be changed with the TimelineTrack and the Target is the MaterialInstance we set in the BP_Goblin_Slingshot and BP_Goblin_Spear. We dissolve for the Mesh and Weapon separately. The Dissolve func in cpp gets called in MulitcastHandleDeath.