ShowingDamageText
In AuraPlayerController the Logic for where and how the DamageText is shown.
In AuraAttributeSet we call the func from the Controller an add the values in PostGameplayEffectExecute. In BP_DamageTextComponent after UpdateDamageText added a delay of 1 second (animation length) and than destroyComponent. Also set the UI->Space->Screen. Added the BP_DamageTextComponent to BP_AuraPlayerController.