DamageCalculationCoefficients
Added a new CurveTable CT_DamageCalculationCoefficients and added the Curves ArmorPenetration and EffectiveArmor as ConstantCurves.
In ExexCalc_Damage getting the CombatInterface for the SourceAvatar and TargetAvatar and the CharacterClassInfo DataTable from UAuraAbilitySystemLibrary.
Accessing the CurveTable in the variable in CharacterClassInfo class and FindCurve by Name. After that Eval the PlayerLevel and set the value. It is done for ArmorPenetration and EffectiveArmor. So instead of hardcoded value we have coefficients based on curve tables.