Damage Types
Added in AuraGameplaytags an  Array of FGameplayTag and added the Tag Damage.Fire in the cpp and added it in the array.
Added AuraDamageGameplayAbiltiy based on AuraGameplayAbility. This class for all DamageTypes. AuraProjectileSpell is now based on AuraDamageGemeplayAbility, so now the UGameplayEffect DamageEffectClass variable is now in the parent with also and newly added TMap of DamageTypes based on the FGameplayTag and stored in an FScalableFloat. So every DamageType can have a CurveTable. So choose the Tag for the Type and a CurveTable. The FScalableFloat Damage in AuraGemplayAbility is now deleted. In the Projectile looping through the map to get the right values associated to it. Also loopin in ExecCalc throug the array to get the base value before any calculation.