Resistance Damage Reduction
Capturing the Resistances. Added a TMap so for every Tag an Capture Definition.
Now Looping through the TMap DamageTypesToResistances in AuraGameplayTags.
Getting the CaptureDef out of every Value for the AttemptCalculateCapturedAttributeMagnitude func for the Resistance variable. Clamping Resistance between 0, 100. Reducing the DamageTypeValue var with the Resistance var. Now the Damage var will be lower.