BlackboardKeys
Added in Blackboard DistanceToTarget float and TargetToFollow type Object and set its BaseClass to Actor. In BTS added variable TargetToFollow_Selector (BlackboardKeySelector Type) and DistanceToTarget_Selector (BlackboardKeySelector Type) and TargetName FName. Added func SetTargetTag. Looking for the controlled Pawn and setting the Tag if the Actor has Player as an Tag (BP_AuraCharacter set the Tag to Player) than Target Tag is set to Enemy, if not set it to Player.