Finding the nearest Player
Added new Func called FindNearestPlayer in BTS_FindNearestPlayer. Setting ClosestActor Type Actor and setting ClosestDistance Type float and an base value like 999999999. Calling then GetAllActorsWithTag and now looping through each element. In the loop setting the each element as local variable called Actor type Actor and checking if it is valid. After that checking if the controlled pawn is valid. Added ControlledPawn as an Input in FindNearestPlayer func in the EventGraph. Parallel getting the distance to the Actor from the Array and setting its value and Name it CalculatedDistance. Branch to true if the CalculatedDistance is less than ClosestDistance and setting the value to ClosestDistance and also setting the Actor to ClosestActor. If the loop is completed calling SetBlackboardValueAsObject and SetBlackboardValueAsFloat. Connecting the selector variable to the key and the values of ClosestActor and ClosestDistance. In BT_EnemyBehaviorTree in Detail->Default connecting the Selectors with the key from the Blackboard. After that to test calling MoveTo as an Task setting the BlackboardKey to TargetToFollow.