AI and EffectActors
Added bools for how the Effects should response on Application and if the enemy could be effected.  In ApplyEffectToTarget and OnOverlap checking if the TargetActor has the Enemy Tag and bApplyEffectsToEnemies is false than return. Both bools are EditAnywhere so can be checked in Editor in the BP_EffectActor children. Deleted all DestroyActor func in the BP of the EffectActors. Instead calling destroy in ApplayEffectToTarget if !bIsInfinte.