Add combat target support to AuraEnemy and interface
Introduced SetCombatTarget and GetCombatTarget methods to the EnemyInterface and implemented them in AAuraEnemy. Added a CombatTarget property to AAuraEnemy, enabling enemies to track and expose their current combat target. Updated related animation and behavior tree assets. In GA_MeleeAttack before the montage playing GetAvatarActorFromActorInfo->GetCombatTarget->GetActorLocation. Also from the ActorInfo Casting to CombatInterface and calling UpdateFacingTarget. Target is the ActorLocation. In BTT_Attack befor TryActivateAbilites setting the SetCombatTarget from the BlackboardKeySelector->TargetToFollow after checking if it the Actor is Valid.