TaggedMontage
Introduced the FTaggedMontage struct to associate animation montages with gameplay tags. Added AttackMontages array to AuraCharacterBase and implemented GetAttackMontages in both the interface and character class, enabling retrieval of tagged attack montages for combat abilities. In GA_MeleeAttack before play montage calling GetAttackMontages getting the ReturnValue and break it open. So Montage goes to the play montage input and the MontageTag goes to WaitGameplayEvent input. In BP_GobinSpear adding an struct element in AttackMontages.