Shaman Attack Montage
Implemented AM_Attack_Shaman withe MotionWarping and Event Track. For MotionWarping edited the WarpTargetName to FacingTarget, unchecked the WarpTranslation to just rotate to the Target and the RotationType to Facing. In Event added AN_MontageEvent and added the Montage.Attack.Weapon Tag. In BP_Shaman added the AnimationMontage and the Tag. For the TipSocket for the Weapon I added and Socket for the Weapon (Staff).