Shaman AttackAbility
Added GA_EnemyFireBolt based on the GA_RangedAttack. Just switched the ProjectileClass to BP_FireBolt and the DamageType is now Damage.Fire and the Value is from CT_Damage->Ability.FireBolt. Also added sounds when the enemy launches his FireBolt to the AnimationMontage and also did it to the Animation Montage of the Slingshot enemy. In the DataAsset DA_CharacterClassInfo added the GA_EnemyFireBolt to the Elementalist Class. Reduced in BP_Shaman the BaseWalkSpeed variable and in the rotation settings the rotation rate to 200°.