Dead Blackboard Key
Added a new Blackboard Key to AuraEnemy in the Die func. SetValueAsBool("Dead").
In the Blackboard also added the Key as Bool and called it Dead. In the BehaviorTree before the enemy decides if it is an Ranged or Melee or Moving to the Target it check alongside with, Do I have a Target and Am I Not HitReacting, Am I Alive?