MeleeImpact GameplayCue
Added GC_MeleeImpact based on GameplayCueNotifyStatic. From the Output Pin break the Parameters. From there you get the information like Location, EffectCauser, SourceObject, EffectContext and more. Then just calling PlaySoundAtLocation(Plugging the location in), GetBloodEffect(SourceObject), SpawnSystemAtLocation(Location and for the SystemTamplate plug in the Return Value from GetBloodEffect). In Ga_MelleAttack calling ExecuteGameplayCueWithParamsOnOwner :) and using the Tag (also implemented for now in Editor and used in the GC) making GameplayCueParamters from the input. From there plugging in the location and get GetAvatarActorFromActorInfo for the EffefctCauser. As SourceObject using the GetCombatTarget function.