Attribute Upgrade Buttons
In WBP_AttributeMenu if the AttributePointsChandedDelegate gets fired I call the new function SetButtonsEnabled. It checks if the AttributePoints integer from the Delegate is greater than 0. If it is it call the new SetButtonEnabled function, from WBP_TextValueButtonRow, where it set the bool from the button, to true for each Row of the MainStats. False otherwise.