Spell Globe Button
For the Spell Menu: created a Spell Globe Button. Just duplicate the ValueButton and deleted some functions the new Widget doesn´t need. As Image_Icon using a Lock Symbol and the Ring is now a button with Ring Brush for all states. Also a Image_Selection as an Halo with animation. Will be used if you select a Spell.
Also changed the naming of the Delegate in OverlayWidgetController for better understanding. Also a typo in the Typename for the delegate. The type is now FOnPlayerStatChandedSignature. The var name is now OnPlayerLevelChangedDelegate.